Move SidebarGroupAnimation into its own file
SidebarWorkspaceGroupingMetrics.swift goes back to holding only spatial
layout constants (Greptile review note). Also drops the now-dead
dropGapHeight: the animated drop gap was replaced by the live reorder.
